requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
dj-database-url==2.2.0
gunicorn==23.0.0
whitenoise==6.8.2
